move some stuff to GrimProcessor
