+++ common/src/main/java/ac/grim/grimac/checks/Check.java
@@ -9,10 +9,6 @@ import ac.grim.grimac.api.storage.verbose.VerboseBuf;
 import ac.grim.grimac.api.storage.verbose.VerboseRenderContext;
 import ac.grim.grimac.internal.storage.verbose.VerboseRegistry;
 import ac.grim.grimac.player.GrimPlayer;
-import com.github.retrooper.packetevents.protocol.packettype.PacketType;
-import com.github.retrooper.packetevents.protocol.packettype.PacketTypeCommon;
-import com.github.retrooper.packetevents.protocol.player.ClientVersion;
-import com.github.retrooper.packetevents.protocol.player.DiggingAction;
 import lombok.Getter;
 import lombok.Setter;
 import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
@@ -22,14 +18,10 @@ import org.jetbrains.annotations.Nullable;
 import java.util.Objects;
 import java.util.function.Supplier;
 
-import static com.github.retrooper.packetevents.wrapper.play.client.WrapperPlayClientPlayerFlying.isFlying;
-
 // Class from https://github.com/Tecnio/AntiCheatBase/blob/master/src/main/java/me/tecnio/anticheat/check/Check.java
 @Getter
 public class Check extends GrimProcessor implements AbstractCheck {
     private static final FlagEvent.Channel FLAG_CHANNEL = GrimAPI.INSTANCE.getEventBus().get(FlagEvent.class);
-
-    protected final @NotNull GrimPlayer player;
 
     // violations
     public double violations;
@@ -60,7 +52,7 @@ public class Check extends GrimProcessor implements AbstractCheck {
     private boolean noModifyPacketPermission;
 
     public Check(final @NotNull GrimPlayer player) {
-        this.player = Objects.requireNonNull(player, "player");
+        super(player, false);
 
         final CheckData checkData = this.getClass().getAnnotation(CheckData.class);
         if (checkData != null) {
@@ -106,8 +98,8 @@ public class Check extends GrimProcessor implements AbstractCheck {
     }
 
     public final void updatePermissions() {
-        if (configName == null) return;
-        final String id = configName.toLowerCase();
+        if (getConfigName() == null) return;
+        final String id = getConfigName().toLowerCase();
         exemptPermission = player.hasPermission("grim.exempt." + id);
         noSetbackPermission = player.hasPermission("grim.nosetback." + id);
         noModifyPacketPermission = player.hasPermission("grim.nomodifypacket." + id);
@@ -231,19 +223,16 @@ public class Check extends GrimProcessor implements AbstractCheck {
 
     @Override
     public final void reload(@NotNull ConfigManager configuration) {
-        if (configName != null) {
-            decay = configuration.getDoubleElse(configName + ".decay", defaultDecay);
-            setbackVL = configuration.getDoubleElse(configName + ".setbackvl", defaultSetbackVL);
-            displayName = configuration.getStringElse(configName + ".displayname", checkName);
-            description = configuration.getStringElse(configName + ".description", defaultDescription);
+        if (getConfigName() != null) {
+            decay = configuration.getDoubleElse(getConfigName() + ".decay", defaultDecay);
+            setbackVL = configuration.getDoubleElse(getConfigName() + ".setbackvl", defaultSetbackVL);
+            displayName = configuration.getStringElse(getConfigName() + ".displayname", checkName);
+            description = configuration.getStringElse(getConfigName() + ".description", defaultDescription);
 
             if (setbackVL == -1) setbackVL = Double.MAX_VALUE;
         }
-        onReload(configuration);
+        super.reload(configuration);
     }
-
-    @Override
-    public void onReload(@NotNull ConfigManager config) {}
 
     public boolean alert(String verbose) {
         return alert(constant(verbose));
@@ -270,53 +259,6 @@ public class Check extends GrimProcessor implements AbstractCheck {
 
     public String formatOffset(double offset) {
         return offset > 0.001 ? String.format("%.5f", offset) : String.format("%.2E", offset);
-    }
-
-    public static boolean isTransaction(PacketTypeCommon packetType) {
-        return packetType == PacketType.Play.Client.PONG ||
-                packetType == PacketType.Play.Client.WINDOW_CONFIRMATION;
-    }
-
-    public static boolean isAsync(PacketTypeCommon packetType) {
-        return packetType == PacketType.Play.Client.KEEP_ALIVE
-                || packetType == PacketType.Play.Client.CHUNK_BATCH_ACK
-                || packetType == PacketType.Play.Client.RESOURCE_PACK_STATUS;
-    }
-
-    public boolean isUpdate(PacketTypeCommon packetType) {
-        return isFlying(packetType)
-                || packetType == PacketType.Play.Client.CLIENT_TICK_END
-                || isTransaction(packetType);
-    }
-
-    public boolean isTickPacket(PacketTypeCommon packetType) {
-        if (isTickPacketIncludingNonMovement(packetType)) {
-            if (isFlying(packetType)) {
-                return !player.packetStateData.lastPacketWasTeleport && !player.packetStateData.lastPacketWasOnePointSeventeenDuplicate;
-            }
-            return true;
-        }
-        return false;
-    }
-
-    public boolean isTickPacketIncludingNonMovement(PacketTypeCommon packetType) {
-        // On 1.21.2+ fall back to the TICK_END packet IF the player did not send a movement packet for their tick
-        // TickTimer checks to see if player did not send a tick end packet before new flying packet is sent
-        if (player.getClientVersion().isNewerThanOrEquals(ClientVersion.V_1_21_2)
-                && !player.packetStateData.didSendMovementBeforeTickEnd) {
-            if (packetType == PacketType.Play.Client.CLIENT_TICK_END) {
-                return true;
-            }
-        }
-
-        return isFlying(packetType);
-    }
-
-    // prevent causing exploits with packet cancelling (ie noslow)
-    public boolean canCancel(DiggingAction action) {
-        return action != DiggingAction.RELEASE_USE_ITEM
-                // we check client version here because 1.8- doesn't predict dropping items, so we can cancel them. (see CompensatedInventory)
-                && (action != DiggingAction.DROP_ITEM && action != DiggingAction.DROP_ITEM_STACK || player.getClientVersion().isOlderThanOrEquals(ClientVersion.V_1_8));
     }
 
     private static @NotNull Supplier<String> constant(String verbose) {
